creating conftest.py for shared functions; updating ner_test to use python api; updating config file to use gpt-4o-mini to be more consistent
