+++ src/tests/ner_test.py
@@ -1,44 +1,38 @@
 """Testing Simple NER examples."""
 
 from pathlib import Path
-import json
+import asyncio
+import yaml
 
 import pytest
-from click.testing import CliRunner
 
-from structsense.cli import cli
+from structsense.app import StructSenseFlow
+from .conftest import skip_if_no_openrouter
 
 pytestmark = pytest.mark.requires_openrouter
 
-CONFIG_PATH = str(Path(__file__).parent / "configs/ner-config_free.yaml")
-ENV_PATH = str(Path(__file__).parent / "configs/.env_example")
+CONFIG_PATH = Path(__file__).parent / "configs/ner-config_extractonly.yaml"
 SOURCE_TEXT = "Retinal ganglion cell (RGC) axons and synapses were genetically labeled via AAV transduction"
 
 
-def test_ner_1(tmp_path):
-    """Test the ENR extraction with a simple text input and a free model."""
-    runner = CliRunner()
-    # uses OPENROUTER_API_KEY set as environment variable (not provided in the env file) for authentication (the model is free)
-    result = runner.invoke(
-        cli,
-        [
-            "extract",
-            "--env_file",
-            ENV_PATH,
-            "--config",
-            CONFIG_PATH,
-            "--source_text",
-            SOURCE_TEXT,
-            "--save_file",
-            str(tmp_path / "enr_result.json"),
-        ],
+@skip_if_no_openrouter
+def test_ner_1():
+    """Test the NER extraction with a simple text input.
+    It only checks if it extracts any entities, since I observed inconsistency.
+    """
+    with open(CONFIG_PATH) as f:
+        config = yaml.safe_load(f)
+
+    flow = StructSenseFlow(
+        agent_config=config["agent_config"],
+        task_config=config["task_config"],
+        embedder_config=config.get("embedder_config", {}),
+        source_text=SOURCE_TEXT,
     )
+    enr_result = asyncio.run(flow.information_extraction_task())
 
-    with open(tmp_path / "enr_result.json", "r") as f:
-        enr_result = json.load(f)
-
-    # testing if we get any entities
-    assert result.exit_code == 0
+    # uses OPENROUTER_API_KEY set as environment variable for authentication
+    assert enr_result is not None
     assert "entities" in enr_result
     assert len(enr_result["entities"]) > 0
     # print the extracted entities for visual inspection (hard to assert exact entities, at least with this model)
